# backend/app/Http/Requests/Auth/UpdatePasswordRequest.php
<?php

namespace App\Http\Requests\Auth;

use Illuminate\Foundation\Http\FormRequest;
use OpenApi\Attributes as OA;

#[OA\Schema(
    schema: 'UpdatePasswordRequest',
    required: ['current_password', 'password', 'password_confirmation'],
    properties: [
        new OA\Property(property: 'current_password', type: 'string', format: 'password', example: 'oldpassword'),
        new OA\Property(property: 'password', type: 'string', format: 'password', example: 'newpassword123'),
        new OA\Property(property: 'password_confirmation', type: 'string', format: 'password', example: 'newpassword123'),
    ]
)]
class UpdatePasswordRequest extends FormRequest
{
    public function authorize(): bool
    {
        return true;
    }

    public function rules(): array
    {
        return [
            'current_password' => ['required', 'string', 'current_password'],
            'password' => ['required', 'string', 'min:8', 'confirmed'],
        ];
    }
}